backend/requirements.txt
anyio==4.9.0
fastapi==0.109.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
priority==2.0.0
pydantic==1.10.22
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.14.0
wsproto==1.2.0
httpx==0.25.2
python-dotenv==1.0.0